% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{bandw}
\alias{bandw}
\title{Bandwidth calculation based on AR(1) approximation}
\usage{
bandw(vhat)
}
\arguments{
\item{vhat}{Estimated variance}
}
\value{
st Bandwidth-corrected variance
}
\description{
Function computes automatic bandwidth based on AR(1) approximation of
estimated vector of vhat with equal weight 1
}
\references{

}
