% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{doorder}
\alias{doorder}
\title{Order estimation procedure}
\usage{
doorder(y, z, x, m, eps, eps1, maxi, fixb, betaini, printd)
}
\arguments{
\item{y}{dependent variables in matrix form}

\item{z}{matrix of independent variables with coefficients are allowed to change across
regimes}

\item{x}{matrix of independent variables with coefficients constant across regimes}

\item{m}{maximum number of breaks}

\item{eps}{convergence criterion for iterative recursive computation}

\item{eps1}{trimming level}

\item{maxi}{maximum number of iterations}

\item{fixb}{option to use fixed initial input \eqn{\beta}. If \code{1},
the model will use values given in \code{betaini}. If \code{0}, betaini is skipped}

\item{betaini}{Initial \eqn{beta_0} to use in estimation}

\item{printd}{option to print results of iterations for partial change model}
}
\value{
A list that contains following:
\item{mBIC}{number of breaks selected by BIC}
\item{mLWZ}{number of breaks selected by LWZ}
}
\description{
The function carry out the procedure to estimate order
 using BIC and the criterion of Liu, Wu and Zidek
}
\references{

}
