% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procedures.R
\name{dotest}
\alias{dotest}
\title{SupF, UDMax & WDMax testing procedure}
\usage{
dotest(
  y,
  z,
  x,
  m,
  eps,
  eps1,
  maxi,
  fixb,
  betaini,
  printd,
  prewhit,
  robust,
  hetdat,
  hetvar
)
}
\arguments{
\item{y}{dependent variables in matrix form}

\item{z}{matrix of independent variables with coefficients are allowed to change across
regimes}

\item{x}{matrix of independent variables with coefficients constant across regimes}

\item{m}{maximum number of breaks}

\item{eps}{convergence criterion for iterative recursive computation}

\item{eps1}{trimming level}

\item{maxi}{maximum number of iterations}

\item{fixb}{option to use fixed initial input \eqn{\beta}. If \code{1},
the model will use values given in \code{betaini}. If \code{0}, betaini is skipped}

\item{betaini}{Initial \eqn{beta_0} to use in estimation}

\item{printd}{option to print results of iterations for partial change model}

\item{prewhit}{option to use AR(1) for prewhitening}

\item{robust, hetdat, hetvar}{options on error terms assumptions}
}
\value{
A list that contains following:
\itemize{
\item{ftest}{SupF test statistics}
\item{cv}{Critical values for Sup F test }
\item{wftest}{Weighted Double Max test}
\item{cvm}{Critical values for Dmax test}
}
}
\description{
The procedure calculate the test statistics and print results of the 2 main tests:
\itemize{
\item{SupF test} {F test of 0 vs m breaks}
\item{Double Max test} {UDMax: the unweighted version
and WDMax: the weighted version}
}
}
