% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getcv1}
\alias{getcv1}
\title{SupF test Critical Values}
\usage{
getcv1(signif, eps1)
}
\arguments{
\item{signif}{significant level}

\item{eps1}{trimming level}
}
\value{
cv Critical value of SupF test
}
\description{

}
