% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getcv2}
\alias{getcv2}
\title{SupF(l+1|l) test Critical Values}
\usage{
getcv2(signif, eps1)
}
\arguments{
\item{signif}{significant level}

\item{eps1}{trimming level}
}
\value{
cv Critical value of SupF(l+1|l) test
}
\description{

}
