% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{kern}
\alias{kern}
\title{Evaluate quadratic kernel of x}
\usage{
kern(x)
}
\arguments{
\item{x}{Value to evaluate kernel}
}
\value{
ker Kernel value of x
}
\description{
Function computes the quadratic kernel by transforming
value of x according to \eqn{k(\delta) = 3\frac{sin(\delta)}{\delta} -
\frac{cos(\delta)}{\delta^2}} where \eqn{\delta(x) = \frac{x6\pi}{5}} is
the transformation of x.
}
