% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{kron}
\alias{kron}
\title{Kronecker Tensor Product}
\usage{
kron(A, B)
}
\arguments{
\item{A}{Matrix A}

\item{B}{Matrix B}
}
\value{
out: Kronecker Tensor Product of A (X) B
}
\description{
Function computes Kronecker Tensor Product between matrix A and matrix B
}
\examples{
A = matrix(c(1,2),1,2)
B = matrix(c(1:4),2,2)
kron(A,B)
}
\author{
Linh Nguyen
}
