% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{OLS}
\alias{OLS}
\title{OLS regression in matrix form}
\usage{
OLS(y, x)
}
\arguments{
\item{y}{matrix of dependent variables}

\item{x}{matrix of independent variables}
}
\value{
b: OLS estimates of the regression
}
\description{
Function computes OLS estimates of the regression y on x
in matrix form, \eqn{\beta = (X'X)^{-1} X'Y}
}
