% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{rot90}
\alias{rot90}
\title{Utilities functions}
\usage{
rot90(A)
}
\arguments{
\item{A}{matrix to be rotated 90 degree counterclockwise}
}
\value{
B: rotated matrix A
}
\description{
Rotate a matrix
}
\details{
Function rotate sa matrix 90 degree counterclockwise
}
\examples{
A = matrix(c(1:10),5,2)
rot90(A)
}
\author{
Linh Nguyen
}
