% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ssr}
\alias{ssr}
\title{SSR computation}
\usage{
ssr(y, zz)
}
\arguments{
\item{y}{matrix of dependent variables}

\item{zz}{matrix of independent variables}
}
\value{
SSR: Sum of Squared Residuals
}
\description{
Function computes Sum of Squared Residuals based on OLS estimates, \eqn{\beta}
}
