% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getdmax}
\alias{getdmax}
\title{Dmax test Critical Values}
\usage{
getdmax(signif, eps1)
}
\arguments{
\item{signif}{significant level}

\item{eps1}{trimming level}
}
\value{
cv Critical value of SupF test
}
\description{

}
