% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ssr}
\alias{ssr}
\title{Recursive SSR computation of a segment}
\usage{
ssr(start, y, z, h, last)
}
\arguments{
\item{start}{starting index of the sample used}

\item{y}{dependent vars}

\item{z}{matrix of time-variant regressors (dimension q)}

\item{h}{minimal length of segment}

\item{last}{ending index of the last segment considered}
}
\value{
out The vector of recursive SSR of length (last-start+1)
}
\description{
Function computes the recursive residuals of period from 'start' to 'last'
}
